backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
numpy==2.3.4
pandas==2.3.3
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
requests==2.32.5
scikit-learn==1.7.2
supabase==2.22.0
gdown==5.1.0